Warning fixes.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@33344 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/mstream.cpp
@@ -54,7 +54,7 @@ wxMemoryInputStream::wxMemoryInputStream(const void *data, size_t len)
 
 wxMemoryInputStream::wxMemoryInputStream(const wxMemoryOutputStream& stream)
 {
-    int len = stream.GetLength();
+    ssize_t len = (ssize_t)stream.GetLength();
     if (len == wxInvalidOffset) {
         m_i_streambuf = NULL;
         m_lasterror = wxSTREAM_EOF;
